+++ lua/lsp/config/pyright.lua
@@ -1,8 +1,4 @@
-local runtime_path = vim.split(package.path, ";")
-table.insert(runtime_path, "lua/?.lua")
-table.insert(runtime_path, "lua/?/init.lua")
-
-local opts = {
+return {
 	cmd = { "pyright-langserver", "--stdio" },
 
 	single_file_support = true,
@@ -16,17 +12,4 @@ local opts = {
 			},
 		},
 	},
-
-	--
-	on_attach = function(client, bufnr)
-		-- 通用配置
-		require("lsp.config.common-attach").commont_on_attach(client, bufnr)
-	end,
-}
-
-return {
-	on_setup = function(server)
-		server:setup(opts)
-	end,
-    opts = opts,
 }