+++ lua/lsp/mason.lua
@@ -1,0 +1,56 @@
+local status, mason = pcall(require, "mason")
+if not status then
+	vim.notify("没有安装插件: mason")
+	return
+end
+
+mason.setup({
+	ui = {
+		icons = {
+			package_installed = "✓",
+			package_pending = "➜",
+			package_uninstalled = "✗",
+		},
+	},
+})
+
+local status, mason_lspconfig = pcall(require, "mason-lspconfig")
+if not status then
+	vim.notify("没有安装插件: mason-lspconfig")
+	return
+end
+
+local servers = {
+	"pyright",
+	"sumneko_lua",
+	"gopls",
+	"clangd",
+}
+
+mason_lspconfig.setup({
+	ensure_installed = servers,
+	automatic_installation = true,
+})
+
+local lspconfig_status_ok, lspconfig = pcall(require, "lspconfig")
+if not lspconfig_status_ok then
+	return
+end
+
+local opts = {}
+
+for _, server in pairs(servers) do
+	opts = {
+		on_attach = require("lsp.handlers").on_attach,
+		capabilities = require("lsp.handlers").capabilities,
+	}
+
+	server = vim.split(server, "@")[1]
+
+	local require_ok, conf_opts = pcall(require, "lsp.config." .. server)
+	if require_ok then
+		opts = vim.tbl_deep_extend("force", conf_opts, opts)
+	end
+
+	lspconfig[server].setup(opts)
+end